PLEASE READ FULL MESSAGE
Okay, apparently php cannot climb up the directory structure, only down.
Thus the images for items have to be in a folder down the directory
structure (in the main project directory). We have to be very careful
what we sync into the project from the remote machine now because we
probably don't want all of the item images added to the github project.