Reformatted to be more correct and appropriate
Dislikes is not a correct term to describe a disliked item. A dislike is
a personal feeling of distaste towards something, but it does not refer
to the something itself. Therefore it should say "disliked items"
instead. Favorites could be considered applicable but it was changed to
be more consistent. The emoticons were removed because they felt out of
tone with the rest of the site which does not make use of them anywhere
else and has a clean, utility-like feeling to it.