Fixed padding and browse query
I fixed the problem with the browse query from before.

Also, we seriously need a padding of 120 on every page at the very
least. I know you don't like how it moves the image down David, but if
we move it up any alerts on the home page make the web page look cheaply
constructed. This is the way it has to be.